+++ Documentation/Doxyfile-pal
@@ -1,7 +1,7 @@
 @INCLUDE = Doxyfile
 PROJECT_NAME = "PAL"
 OUTPUT_DIRECTORY = _build/doxygen-pal
-INPUT = ../Pal/include/pal ../Pal/src
+INPUT = ../Pal/include/arch/x86_64 ../Pal/include/pal ../Pal/src
 EXCLUDE = ../Pal/src/host
 MACRO_EXPANSION = YES
 EXPAND_ONLY_PREDEF = YES
