+++ Pal/src/Makefile
@@ -21,7 +21,7 @@ OBJ_DIR = $(HOST_DIR)/.obj
 include host/$(PAL_HOST)/Makefile.am
 
 CFLAGS += -I. -I../include -I../include/pal -I../include/$(HOST_DIR) \
-    -I$(HOST_DIR) -I../include/lib
+    -I$(HOST_DIR) -I../include/lib -I../include/arch/$(ARCH)
 
 PAL_HOST_MACRO = $(shell echo $(PAL_HOST) | tr '[:lower:]' '[:upper:]' | tr '-' '_')
