+++ Pal/src/do-rel.h
@@ -24,7 +24,7 @@
 
 #include <pal_rtld.h>
 
-#include "dl-machine-x86_64.h"
+#include "dl-machine.h"
 
 #define elf_dynamic_do_rel       elf_dynamic_do_rela
 #define RELCOUNT_IDX             VERSYMIDX(DT_RELACOUNT)
