+++ Pal/src/host/Linux/Makefile
@@ -2,6 +2,7 @@ include ../../../../Scripts/Makefile.configs
 include Makefile.am
 
 CFLAGS	+= -I. -Iinclude -I../.. -I../../../include -I../../../include/pal \
+	-I../../../include/arch/$(ARCH) \
 	-I../../../include/host/Linux -I../../../include/lib \
 	-I../../../include/host/Linux-common
 ASFLAGS += -I. -Iinclude -I../.. -I../../../include
